+++ assembler/src/common/assembly_graph/paths/path_utils.hpp
@@ -111,7 +111,7 @@ namespace debruijn_graph {
     template<class Graph>
     Sequence MergeSequences(const Graph &g, const std::vector<typename Graph::EdgeId> &continuous_path) {
         std::vector<Sequence> path_sequences;
-        std::vector<uint32_t> overlaps;
+        std::vector<size_t> overlaps;
         if (continuous_path.size() == 0)
             return Sequence();
         path_sequences.push_back(g.EdgeNucls(continuous_path[0]));
@@ -119,7 +119,7 @@ namespace debruijn_graph {
             auto end = g.EdgeEnd(continuous_path[i - 1]);
             auto start = g.EdgeStart(continuous_path[i]);
             VERIFY(start == end);
-            uint32_t overlap = g.link_length(start, continuous_path[i - 1], continuous_path[i]);
+            size_t overlap = g.link_length(start, continuous_path[i - 1], continuous_path[i]);
             overlaps.push_back(overlap);
             path_sequences.push_back(g.EdgeNucls(continuous_path[i]));
         }
